feat: 신발장 폴더 추가, 변수/타입 추가
