+++ src/types/shoerack.d.ts
@@ -1,0 +1,19 @@
+export type TShoeInfo = {
+  productId: string;
+  modelNo?: string;
+  modelName: string;
+  brand?: string;
+  link?: string;
+  image: string;
+};
+
+export type TShoeData = {
+  star: number;
+  length: string;
+  width: string;
+  height: string;
+  cushion: string;
+  weight: string;
+  size: string;
+  review: string;
+};
